backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ── Pydantic — pre-built binary wheels for Python 3.11 ───────────────────────
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# ── Email validation (required by pydantic EmailStr) ─────────────────────────
email-validator==2.2.0

# ── HTTP & Utilities ──────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
aiofiles==24.1.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Firebase ──────────────────────────────────────────────────────────────────
firebase-admin==6.6.0
